release 0

+++ src/biz/gelicon/gta/server/data/Team.java
@@ -34,10 +34,10 @@ public class Team  implements Serializable {
 	private Set<Person> persons = new HashSet<>();
 	private Date createDate;
 	private Integer limit;
-	private Integer workedOfDay;
-	private Integer workedOfWeek;
-	private Integer workedOfMonth;
-	private Integer workedOfBeginProject;
+	private Double workedOfDay;
+	private Double workedOfWeek;
+	private Double workedOfMonth;
+	private Double workedOfBeginProject;
 	
 	@Id
 	@javax.persistence.SequenceGenerator(name="newRec", sequenceName="NEWRECORDID")	
@@ -84,35 +84,35 @@ public class Team  implements Serializable {
 	}
 
 	@Transient
-	public Integer getWorkedOfDay() {
+	public Double getWorkedOfDay() {
 		return workedOfDay;
 	}
-	public void setWorkedOfDay(Integer workedOfDay) {
+	public void setWorkedOfDay(Double workedOfDay) {
 		this.workedOfDay = workedOfDay;
 	}
 	
 	@Transient
-	public Integer getWorkedOfWeek() {
+	public Double getWorkedOfWeek() {
 		return workedOfWeek;
 	}
-	public void setWorkedOfWeek(Integer workedOfWeek) {
+	public void setWorkedOfWeek(Double workedOfWeek) {
 		this.workedOfWeek = workedOfWeek;
 	}
 	
 	@Transient
-	public Integer getWorkedOfMonth() {
+	public Double getWorkedOfMonth() {
 		return workedOfMonth;
 	}
 	
-	public void setWorkedOfMonth(Integer workedOfMonth) {
+	public void setWorkedOfMonth(Double workedOfMonth) {
 		this.workedOfMonth = workedOfMonth;
 	}
 	
 	@Transient
-	public Integer getWorkedOfBeginProject() {
+	public Double getWorkedOfBeginProject() {
 		return workedOfBeginProject;
 	}
-	public void setWorkedOfBeginProject(Integer workedOfBeginProject) {
+	public void setWorkedOfBeginProject(Double workedOfBeginProject) {
 		this.workedOfBeginProject = workedOfBeginProject;
 	}
